Super select WIP, Normal name switch
