+++ src/lib-components/index.ts
@@ -12,5 +12,4 @@ export { default as FTableRow } from './FTableRow.vue'
 export { default as FTableCol } from './FTableCol.vue'
 export { default as FTableHeading } from './FTableHeading.vue'
 export { default as FTextarea } from './FTextarea.vue'
-export { default as FBS } from './FBS.vue'
